Support both cjs require and es6 imports for jsonnet library
