+++ JS/jsonnet/read-file.js
@@ -1,5 +1,8 @@
-import fs from "fs";
+// import fs from "fs";
+const fs = require("fs");
 
-export function read_file(path) {
+function read_file(path) {
     return fs.readFileSync(path, { encoding: "utf8" });
 }
+
+module.exports = { read_file }
