+++ JS/jsonnet/src/index.d.mts
@@ -1,0 +1,10 @@
+declare class Jsonnet {
+    constructor();
+    evaluateSnippet(snippet: string): string;
+    destroy(): void;
+    extString(key: string, value: string): this;
+    evaluateFile(filename: string): string;
+    javascriptCallback(name: string, func: Function): this;
+}
+
+export default Jsonnet;
