+++ JS/jsonnet/src/index.mjs
@@ -1,0 +1,3 @@
+import Jsonnet  from "./jsonnet.js";
+
+export default Jsonnet;
