+++ JS/jsonnet/src/jsonnet_wasm_bg.wasm.d.ts
@@ -3,14 +3,7 @@
 export const memory: WebAssembly.Memory;
 export function jsonnet_make(): number;
 export function jsonnet_destroy(a: number): void;
-export function jsonnet_evaluate_snippet(
-    a: number,
-    b: number,
-    c: number,
-    d: number,
-    e: number,
-    f: number
-): void;
+export function jsonnet_evaluate_snippet(a: number, b: number, c: number, d: number, e: number, f: number): void;
 export function jsonnet_evaluate_file(a: number, b: number, c: number, d: number): void;
 export function ext_string(a: number, b: number, c: number, d: number, e: number): void;
 export function get_func(a: number, b: number): number;
