+++ JS/jsonnet/src/snippets/arakoo-jsonnet-17c737407ebd2d3c/read-file.js
@@ -1,5 +1,8 @@
-import fs from "fs";
+// import fs from "fs";
+const fs = require("fs");
 
-export function read_file(path) {
+function read_file(path) {
     return fs.readFileSync(path, { encoding: "utf8" });
 }
+
+module.exports = { read_file }
